pyproject.toml
[project]
name = "tsfm-portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "boto3>=1.42.63",
    "chronos-forecasting>=2.2.2",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "sagemaker>=3.0.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "transformers>=4.57.6",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "ruff>=0.15.5",
]
